other sensors data

+++ Tools/ColorProgressBar.h
@@ -6,31 +6,31 @@
 
 class ColorProgressBar : public QWidget {
     Q_OBJECT
-    Q_PROPERTY(int value READ value WRITE setValue)
-    Q_PROPERTY(int maximum READ maximum WRITE setMaximum)
+    Q_PROPERTY(double value READ value WRITE setValue)
+    Q_PROPERTY(double maximum READ maximum WRITE setMaximum)
     Q_PROPERTY(QColor color READ color WRITE setColor)
 
 public:
     explicit ColorProgressBar(QWidget *parent = nullptr);
 
-    int value() const { return m_value; }
-    int minimum() const { return m_minimum; }
-    int maximum() const { return m_maximum; }
+    double value() const { return m_value; }
+    double minimum() const { return m_minimum; }
+    double maximum() const { return m_maximum; }
     QColor color() const { return m_color; }
 
 public slots:
-    void setValue(int value);
-    void setMinimum(int minimum);
-    void setMaximum(int maximum);
+    void setValue(double value);
+    void setMinimum(double minimum);
+    void setMaximum(double maximum);
     void setColor(const QColor &color);
 
 protected:
     void paintEvent(QPaintEvent *) override;
 
 private:
-    int m_value;
-    int m_minimum = 0;   // Добавлено минимальное значение
-    int m_maximum = 1000;
+    double m_value;
+    double m_minimum = 0;   // Добавлено минимальное значение
+    double m_maximum = 1000;
     QColor m_color = Qt::blue;
 };
 #endif // COLORPROGRESSBAR_H
